Dont allow frontend code to import server code indirectly
