+++ components/LiveControls.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useState } from "preact/hooks";
 import { Page, Site } from "$live/types.ts";
-import { DomInspector } from "$live/inspect.ts";
+import { DomInspector } from "https://deno.land/x/inspect_vscode@0.2.0/mod.ts";
 import { ViewfinderCircleIcon } from "https://esm.sh/@heroicons/react@2.0.12/24/outline?alias=react:preact/compat&external=preact/compat";
 
 declare global {
