+++ mod.ts
@@ -1,17 +1,5 @@
-import LivePage from "$live/components/LivePage.tsx";
-import LiveSections from "$live/components/LiveSections.tsx";
-import LiveControls from "$live/components/LiveControls.tsx";
-import CoreWebVitals from "$live/components/CoreWebVitals.tsx";
 import { live, withLive } from "$live/live.ts";
 import stylesPlugin from "$live/plugins/styles.ts";
 export * from "$live/types.ts";
 
-export {
-  CoreWebVitals,
-  live,
-  LiveControls,
-  LivePage,
-  LiveSections,
-  stylesPlugin,
-  withLive,
-};
+export { live, stylesPlugin, withLive };
