+++ types.ts
@@ -6,8 +6,12 @@ import type { Manifest } from "$fresh/server.ts";
 import type { JSONSchema7 } from "https://esm.sh/v92/@types/json-schema@7.0.11/X-YS9yZWFjdDpwcmVhY3QvY29tcGF0CmQvcHJlYWN0QDEwLjEwLjY/index.d.ts";
 import { LoaderFunction } from "$live/std/types.ts";
 import { createServerTimings } from "$live/utils/timings.ts";
-
-export type { Node } from "$live/workbench.ts";
+export interface Node {
+  label: string;
+  fullPath: string;
+  editLink?: string;
+  children?: Node[];
+}
 
 export type Schema = JSONSchema7;
 
@@ -39,6 +43,7 @@ export interface LiveOptions {
   siteId: number;
   loginUrl?: string;
   domains?: string[];
+  inspectPath?: string;
 }
 
 export interface PageSection {
